Refactor estimators: rename base.py, move configs, fix Flow betas
- Rename estimators/base.py → estimators/stepwise_base.py (name now
  reflects its content: StepwiseEstimator and LossBasedEstimator)
- Move OptimizerConfig and InferenceConfig out of stepwise_base into
  gaussian.py where they actually belong; stepwise_base uses Any for
  the duck-typed optimizer/inference config params
- Add betas=(0.9, 0.9) to flow.py OptimizerConfig and wire into AdamW
  (Flow was silently using PyTorch default beta2=0.999 instead of the
  lower beta2=0.9 needed for non-stationary SBI training data)
- Remove OptimizerConfig/InferenceConfig from estimators/__init__ public
  API (each estimator now owns its own)
