+++ falcon/estimators/gaussian.py
@@ -26,12 +26,29 @@ from omegaconf import OmegaConf
 
 from falcon.estimators.networks import build_mlp
 from falcon.priors.product import TransformedPrior
-from falcon.estimators.base import (
+from falcon.estimators.stepwise_base import (
     LossBasedEstimator,
     TrainingLoopConfig,
-    OptimizerConfig,
-    InferenceConfig,
 )
+
+
+@dataclass
+class OptimizerConfig:
+    """Optimizer and scheduler parameters."""
+
+    lr: float = 1e-3
+    betas: tuple = (0.9, 0.9)  # Lower beta2 for dynamic SBI setting
+    lr_decay_factor: float = 0.1
+    scheduler_patience: int = 8
+
+
+@dataclass
+class InferenceConfig:
+    """Inference and sampling parameters."""
+
+    gamma: float = 0.5
+    discard_samples: bool = False
+    log_ratio_threshold: float = -20.0
 from falcon.core.logger import log, debug, info, warning, error
